+++ lib/src/features/nutrition/features/Mealview/screen/mealview.dart
@@ -18,11 +18,11 @@ class Mealview extends StatelessWidget {
   });
   final String foodName;
   final String foodImage;
-  final int calories;
-  final int weight;
-  final int fat;
-  final int carbs;
-  final int protein;
+  final double calories;
+  final double weight;
+  final double fat;
+  final double carbs;
+  final double protein;
   final List<String> Ingredients;
   final List<String> steps;
   final List<Map<String, dynamic>> tips;
@@ -380,7 +380,7 @@ class NutrientCard extends StatelessWidget {
   final IconData icon;
   final Color iconColor;
   final String title;
-  final int content;
+  final double content;
   final String unit;
 
   const NutrientCard({
